Added maket open check to BaseStrat
